% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_man.R
\name{vec_sample}
\alias{vec_sample}
\title{vec_sample}
\usage{
vec_sample(x, size, replace = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{size}{Number of elements to  be returned as sample}

\item{replace}{a element will repeat in the sample, by default it will be FALSE.}
}
\description{
Samples a specified number of elements from a vector.
}
