% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_arith.R
\name{vec_sort}
\alias{vec_sort}
\title{vec_sort}
\usage{
vec_sort(x, decreasing = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{decreasing}{if TRUE then the vector will be in descending order.
#' @example
 /dontrun vec_sort(c(3,1,4,5,2))}
}
\description{
Sorts a vector in ascending order.
}
