% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_man.R
\name{vec_subset}
\alias{vec_subset}
\title{vec_subset}
\usage{
vec_subset(x, condition)
}
\arguments{
\item{x}{A numeric vector.}

\item{condition}{any logical condition.}
}
\description{
Subsets the vector on given logical condition and returns the subset.
}
