% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_arith.R
\name{vec_uniq}
\alias{vec_uniq}
\title{vec_uniq}
\usage{
vec_uniq(x)
}
\arguments{
\item{x}{A numeric vector.
#' @example
 /dontrun vec_uniq(c(5,5,4,3,2,2,1))}
}
\description{
Removes all the duplicate values in a vector.
}
